+++ CHANGELOG.md
@@ -63,6 +63,7 @@ without arguments starts a full interactive Python console.
 - EvTable should now correctly handle columns with wider asian-characters in them.
 - Update Twisted requirement to >=2.3.0 to close security vulnerability
 - Add `$random` inlinefunc, supports minval,maxval arguments that can be ints and floats.
+- Add `evennia.utils.inlinefuncs.raw(<str>)` as a helper to escape inlinefuncs in a string.
 
 
 ## Evennia 0.9 (2018-2019)